remove sqlalchemy models
this commit removes the sqlalchemy models from the seamm dashboard package. Models are now in the seamm_datastore package.
