--- seamm_dashboard/models/__init__.py
@@ -1,19 +1,0 @@
-from .models import (  # noqa: F401
-    Job,
-    Flowchart,
-    Project,
-    User,
-    Group,
-    Role,
-    UserProjectAssociation,
-    GroupProjectAssociation,
-)
-from .models import (  # noqa: F401
-    JobSchema,
-    FlowchartSchema,
-    ProjectSchema,
-    UserSchema,
-    GroupSchema,
-    RoleSchema,
-)
-from .import_jobs import import_jobs  # noqa: F401
