+++ seamm_dashboard/routes/api/jobs.py
@@ -19,8 +19,8 @@ from flask import send_from_directory, Response
 from flask_jwt_extended import jwt_required
 
 from seamm_dashboard import db, datastore, authorize, options
-from seamm_dashboard.models import User, Project, Job, JobSchema, Flowchart
-from seamm_dashboard.models import FlowchartSchema
+from seamm_datastore.models import User, Project, Job, Flowchart
+from seamm_datastore.schema import JobSchema, FlowchartSchema
 
 logger = logging.getLogger("__file__")
