+++ seamm_dashboard/routes/api/users.py
@@ -7,14 +7,9 @@ from flask import Response
 from flask_jwt_extended import jwt_required
 
 from seamm_dashboard import db, authorize
-from seamm_dashboard.models import (
-    User,
-    UserSchema,
-    Role,
-    RoleSchema,
-    Group,
-    GroupSchema,
-)
+from seamm_datastore.models import User, Role, Group
+
+from seamm_datastore.schema import UserSchema, RoleSchema, GroupSchema
 
 __all__ = ["add_user", "get_users"]
