+++ seamm_dashboard/routes/auth/views.py
@@ -14,7 +14,8 @@ from flask_jwt_extended import (
 from .forms import LoginForm, ConfirmLogin, UpdateAccountInfoForm
 
 from seamm_dashboard import authorize, db
-from seamm_dashboard.models import User, UserSchema, UserProjectAssociation
+from seamm_datastore.models import User, UserProjectAssociation
+from seamm_datastore.schema import UserSchema
 
 from . import auth
