+++ seamm_dashboard/routes/jobs/views.py
@@ -4,16 +4,16 @@ from flask import request, render_template, flash, redirect, url_for
 
 from flask_jwt_extended import jwt_required, get_current_user
 
-from seamm_dashboard.models import Project, Job
+from seamm_datastore.models import Project, Job
 from seamm_dashboard import db, authorize
 
-from seamm_dashboard.models.util import process_job
-from seamm_dashboard.models.import_jobs import add_job
+from seamm_dashboard.util.util import process_job
+from seamm_dashboard.util.import_jobs import add_job
 
 from . import jobs
 
 from seamm_dashboard.routes.jobs.forms import EditJob, ImportJob
-from seamm_dashboard.models import Job
+from seamm_datastore.models import Job
 
 from werkzeug.utils import secure_filename
