+++ seamm_dashboard/util/__init__.py
@@ -1,0 +1,2 @@
+
+from .import_jobs import import_jobs  # noqa: F401
