+++ seamm_dashboard/util/util.py
@@ -12,12 +12,27 @@ import logging
 from pathlib import Path
 import os
 
-from . import User, Group, Role
-from seamm_dashboard import db
+from seamm_datastore.models import User, Group, Role
+from seamm_dashboard import db, jwt
 
 logger = logging.getLogger(__name__)
 
 time_format = "%Y-%m-%d %H:%M:%S %Z"
+
+@jwt.user_lookup_loader
+def user_loader_callback(jwt_header, jwt_payload):
+    """Function for app, to return user object"""
+
+    if jwt_header:
+        from seamm_datastore.models import User
+
+        username = jwt_payload["sub"]["username"]
+        user = User.query.filter_by(username=username).one_or_none()
+
+        return user
+    else:
+        # return None / null
+        return None
 
 
 def process_flowchart(flowchart_path):
